Move util method to VaadinSession
